Move interpreters to mod
