+++ radix-engine/src/engine/interpreters/mod.rs
@@ -1,4 +1,7 @@
+mod native_interpreter;
 mod scrypto_interpreter;
+mod wasm_runtime;
 
-pub use crate::engine::wasm_runtime::*;
+pub use native_interpreter::*;
 pub use scrypto_interpreter::*;
+pub use wasm_runtime::*;
