+++ radix-engine/src/engine/mod.rs
@@ -4,14 +4,12 @@ mod errors;
 mod interpreters;
 mod kernel;
 mod modules;
-mod native_interpreter;
 mod node;
 mod node_properties;
 mod node_ref;
 mod system_api;
 mod track;
 mod track_support;
-mod wasm_runtime;
 
 pub use actor::*;
 pub use call_frame::CallFrame;
@@ -19,11 +17,9 @@ pub use errors::*;
 pub use interpreters::*;
 pub use kernel::*;
 pub use modules::*;
-pub use native_interpreter::NativeInterpreter;
 pub use node::*;
 pub use node_properties::*;
 pub use node_ref::*;
 pub use system_api::{LockFlags, SystemApi};
 pub use track::*;
 pub use track_support::*;
-pub use wasm_runtime::*;
